refactor: use generated `BaseLanguageServer`
This commit refactors the `LanguageServer` class to be based on the new
`BaseLanguageServer`. The server has also been moved into the
`pygls.lsp` module in an attempt to be more consistent with the client
side code.

It also takes the opportunity to tidy a few things up

- Pushing generic server features such as error reporting and feature
  registration down into the base `Server` class
- Removing LSP specific quirks from the base class e.g.
  `TextDocumentSyncKind`
- Renames `server.lsp` to `server.protocol` to align better with the
  client code
